Add initial UI structure, dependencies
