+++ ui/package.json
@@ -4,9 +4,11 @@
   "private": true,
   "dependencies": {
     "axios": "^0.18.0",
+    "bootstrap": "^4.1.3",
     "react": "^16.5.1",
     "react-dom": "^16.5.1",
     "react-scripts": "1.1.5",
+    "reactstrap": "^6.5.0",
     "sockjs-client": "^1.3.0",
     "stomp-websocket": "^2.3.4-next"
   },
